+++ brewplus/brewplus-ifdb/src/main/java/jmash/Ricetta.java
@@ -2421,13 +2421,13 @@ public class Ricetta extends javax.swing.JInternalFrame {
 		this.unitaMisura = unitaMisura;
 	}
 
-	private int efficienza = Main.config.getEfficienza();
+	private double efficienza = Main.config.getEfficienza();
 
-	public int getEfficienza() {
+	public double getEfficienza() {
 		return this.efficienza;
 	}
 
-	public void setEfficienza(int v) {
+	public void setEfficienza(double v) {
 		this.efficienza = v;
 	}
